Refactor the asdf_maker subpackage
This moves the wcs specific stuff out into its own file and adds an extra test
