+++ dkist/asdf_maker/__init__.py
@@ -1,2 +1,4 @@
+"""
+Functions for building asdf files from a set of calibrated DKIST FITS files and their headers.
+"""
 from .generator import *
-from .helpers import *
